Fixed the image deletion error when url is empty

+++ routes/user_route.js
@@ -126,28 +126,32 @@ router.put(
       return res.status(400).json({ message: "User is not updated" });
     }
 
-    //if user is updated it will also update the avatar
-    //if only user name is updated there will be duplicate entry of the avatar
-    //to delete the older entry,
-    //we are split the url in "/" and it devide url in to 8 array element
-    //7th position is the image name
-    const image_name = avatar_url.split("/")[7];
+    //if avatar_url is not empty we'll delete it
+    //to remove old entries
+    if (avatar_url != "") {
+      //if user is updated it will also update the avatar
+      //if only user name is updated there will be duplicate entry of the avatar
+      //to delete the older entry,
+      //we are split the url in "/" and it devide url in to 8 array element
+      //7th position is the image name
+      const image_name = avatar_url.split("/")[7];
 
-    //Here we are finding path of the image
-    //using it we can delete it
-    const image_path = path.join(
-      __dirname,
-      "../",
-      `/public/uploads/${image_name}`
-    );
+      //Here we are finding path of the image
+      //using it we can delete it
+      const image_path = path.join(
+        __dirname,
+        "../",
+        `/public/uploads/${image_name}`
+      );
 
-    //we are trying to delete the older entry
-    //by using fs.unlink()
-    try {
-      fs.unlinkSync(image_path);
-      //file removed
-    } catch (err) {
-      return res.status(400).json({ message: "Old avatar not deleted" });
+      //we are trying to delete the older entry
+      //by using fs.unlink()
+      try {
+        fs.unlinkSync(image_path);
+        //file removed
+      } catch (err) {
+        return res.status(400).json({ message: "Old avatar not deleted" });
+      }
     }
     return res.status(200).json({ message: "User updated" });
   }
